requirements.txt
et_xmlfile==2.0.0
greenlet==3.1.1
numpy==1.24.2
openpyxl==3.1.1
pandas==1.5.3
psycopg2-binary==2.9.10
pyodbc==4.0.35
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
XlsxWriter==3.0.8
